% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstruct.R
\docType{package}
\name{fstruct}
\alias{fstruct}
\title{fstruct: A package to measure variability in population structure estimates}
\description{
An Fst-based tool to compare variability between Q matrices
with rows of individual ancestry coefficients (the default output of
population structure inference programs such as \href{https://web.stanford.edu/group/pritchardlab/structure.html}{STRUCTURE}
and \href{http://dalexander.github.io/admixture/index.html}{ADMIXTURE}.).
Included functions plot Q matrices using ggplot2, calculate
Fst/Fst-max (a normalized measure of variability) for a
Q matrix, and generate bootstrap replicates of one or more Q
matrices along with associated statistics. Accompanies (insert paper citation
here).
}
\section{fstruct functions}{

The foo functions ...
}

