% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstruct_functions.R
\name{plot.structure}
\alias{plot.structure}
\title{Plot a Q matrix using \code{\link[ggplot2]{ggplot2}}}
\usage{
\method{plot}{structure}(Q, k)
}
\arguments{
\item{Q}{A dataframe, matrix, or array representing a Q matrix. Each row
represents an individual and the last \code{k} columns contain individual
membership coefficients. The first few columns may contain information not
relevant to this plot; their inclusion is optional. When restricted to the
last \code{k} columns, the rows of this matrix should sum to approximately
1.}

\item{k}{The number of ancestral clusters in the Q matrix. Each individual
should have \code{k} membership coefficients. The default color scheme is
"spectral" from \code{\link[RColorBrewer]{RColorBrewer}}, which tolerates
\code{k} values of 11 or fewer.}
}
\value{
A \code{ggplot} object describing a bar plot of membership
  coefficients from the Q matrix.
}
\description{
This function enables graphical visualization of a Q matrix, the default
output of population structure inference software programs such as
\href{https://web.stanford.edu/group/pritchardlab/structure.html}{STRUCTURE}
and \href{http://dalexander.github.io/admixture/index.html}{ADMIXTURE}. In
the output plot, each vertical bar represents a single individual's ancestry;
the height of each color in the bar corresponds to the individual membership
coefficients given by the Q matrix. Because this function produces a
\code{\link[ggplot2]{ggplot}} object, its output can be modified using
standard \code{\link[ggplot2]{ggplot2}} syntax. For a more comprehensive
population structure visualization program, see the program
\emph{\href{https://rosenberglab.stanford.edu/distruct.html}{distruct}}.
}
\examples{
structure.plot(
# Make an example matrix of membership coefficients.
# Each row is an individual. Rows sum to 1.
Q = matrix(c(.4,.2,.4,
             .5,.3,.2,
             .5,.4,.1,
             .6,.1,.3,
             .6,.3,.1),
           nrow = 5,
           byrow = TRUE),
# How many ancestry coefficients per individual?
k = 3
) +
# Below are example, optional modifications to the default plot
  ggtitle("Population A") +
  scale_fill_brewer("Blues") +
  xlab("Individuals")

}
