% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSTruct_functions.R
\name{Q_plot}
\alias{Q_plot}
\title{Plot a Q matrix using ggplot2}
\usage{
Q_plot(Q, K)
}
\arguments{
\item{Q}{A dataframe, matrix, or array representing a Q matrix. Each row
represents an individual and the last \code{K} columns contain individual
membership coefficients. The first few columns may contain information not
relevant to this plot; their inclusion is optional. When restricted to the
last \code{K} columns, the rows of this matrix should sum to approximately
1.}

\item{K}{The number of ancestral clusters in the Q matrix. Each individual
should have \code{K} membership coefficients. The default color scheme is
"spectral" from RColorBrewer, which tolerates
\code{K} values of 11 or fewer.}
}
\value{
A ggplot object describing a bar plot of membership
  coefficients from the Q matrix.
}
\description{
This function enables graphical visualization of a Q matrix, the default
output of population structure inference software programs such as
\href{https://web.stanford.edu/group/pritchardlab/structure.html}{STRUCTURE}
and \href{http://dalexander.github.io/admixture/index.html}{ADMIXTURE}. In
the output plot, each vertical bar represents a single individual's ancestry;
the height of each color in the bar corresponds to the individual membership
coefficients given by the Q matrix. Because this function produces a
ggplot object, its output can be modified using
standard ggplot2 syntax. For a more comprehensive
population structure visualization program, see the program
\emph{\href{https://rosenberglab.stanford.edu/distruct.html}{distruct}}.
}
\examples{
Q_plot(
# Make an example matrix of membership coefficients.
# Each row is an individual. Rows sum to 1.
Q = matrix(c(.4,.2,.4,
             .5,.3,.2,
             .5,.4,.1,
             .6,.1,.3,
             .6,.3,.1),
           nrow = 5,
           byrow = TRUE),
           K = 3 # How many ancestry coefficients per individual?
) +
# Below are example, optional modifications to the default plot
  ggplot2::ggtitle("Population A") +
  ggplot2::scale_fill_brewer("Blues") +
  ggplot2::xlab("Individuals")

}
