% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstruct.R
\docType{package}
\name{fstruct}
\alias{fstruct}
\title{fstruct: A package to measure variability in population structure estimates}
\description{
An \eqn{$F_{ST}$}-based tool to compare variability between \eqn{Q} matrices
with rows of individual ancestry coefficients (the default output of
population structure inference programs such as STRUCTURE and ADMIXTURE).
Included functions plot \eqn{Q} matrices using ggplot2, calculate
\eqn{F_{ST}/F_{ST}^\text{max}}(a normalized measure of variability) for a
\eqn{Q} matrix, and generate bootstrap replicates of one or more \eqn{Q}
matrices along with associated statistics. Accompanies (insert paper citation
here).
}
\section{fstruct functions}{

The foo functions ...
}

