% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSTruct_functions.R
\name{Q_bootstrap}
\alias{Q_bootstrap}
\title{Generate and analyze bootstrap replicates of one or more Q matrices}
\usage{
Q_bootstrap(matrices, n_replicates, K)
}
\arguments{
\item{matrices}{A dataframe, matrix, or array representing a Q matrix, or a (possibly named) list of arbitrarily many such objects. For each Q matrix, matrix rows represents an individual and the last \code{K} columns contain individual membership coefficients (when restricted to the last \code{K} columns, the rows should sum to approximately 1). If the list of matrices is not named, the matrices are numbered and the numbers are used in place of names.}

\item{n_replicates}{The number of bootstrap replicate matrices to generate for each provided Q matrix.}

\item{K}{The number of ancestral clusters in the Q matrix. Each individual should have \code{K} membership coefficients.}
}
\value{
A named list of containing the following entries:
\itemize{
\item \code{bootstrap_replicates}: A named list of lists. Each element is named for a Q matrix provided in \code{matrices} and contains a list of \code{n_replicates} bootstrap replicates of the provided matrix. E.g., if \code{n_replicates = 100} and the first Q matrix in \code{matrices} is named \code{A}, then the first element of \code{bootstrap_replicates}, \code{bootstrap_replicates$bootstrap_matrices_A}, is itself a list of 100 matrices, each representing a bootstrap replicate of matrix A.
\item \code{statistics}: A dataframe containing the output of \code{Q_stat}: \code{Fst}, \code{FstMax}, and \code{ratio} (Fst/FstMax), computed for each bootstrap replicate matrix in \code{bootstrap_replicates}. The ratio Fst/FstMax quantifies the variability of each  Q matrix. The first column, titled \code{Matrix}, is a factor indicating which provided Q matrix the row corresponds to (the matrix name if \code{matrices} is a named list, or a number otherwise). The row names are of the form \code{stats_matrix.replicate} where \code{matrix} is the name of one of the provided Q matrices (or the entry number if the list elements were not named) and replicate is the number of bootstrap replicate (rep takes values from 1 to \code{n_replicates}).
\item \code{plot_boxplot}: A ggplot2 box plot depicting the bootstrap distribution of Fst/FstMax for each matrix in \code{matrices}.
\item \code{plot_violin}: A ggplot2 violin plot depicting the bootstrap distribution of Fst/FstMax for each matrix in \code{matrices}.
\item \code{plot_ecdf}: A ggplot2 empirical cumulative distribution function plot depicting the bootstrap distribution of Fst/FstMax for each matrix in \code{matrices}.
\item \code{test_kruskal_wallis}: Results of a Kruskal-Wallis test performed on the bootstrap distributions of Fst/FstMax. This test is a non-parametric statistical test of whether all provided bootstrap distributions are identically distributed.
\item \code{test_pairwise_wilcox}: Results of a Wilcoxon rank-sum test performed on the bootstrap distributions of Fst/FstMax. This test is a non-parameteric statistical test of whether \emph{each pairwise combination} of provided bootstrap distributions is identically distributed.
}
}
\description{
Generates bootstrap replicate Q matrices, computes Fst/FstMax for each bootstrap replicate, produces several plots of the bootstrap distributions of Fst/FstMax for each provided Q matrix, and runs two statistical tests comparing these bootstrap distributions. The tests comparing bootstrap distributions of Fst/FstMax facilitate statistical comparison of the variability in each of multiple Q matrices.
}
\examples{
# Use Q_simulate to generate 4 random Q matrices
A = Q_simulate(alpha = .1,
               lambda = c(.5, .5),
               rep = 1,
               popsize = 20,
               seed = 1)

B = Q_simulate(alpha = .1,
               lambda = c(.5, .5),
               rep = 1,
               popsize = 20,
               seed = 2)

C = Q_simulate(alpha = 1,
               lambda = c(.5, .5),
               rep = 1,
               popsize = 20,
               seed = 3)

D = Q_simulate(alpha = 1,
               lambda = c(.5, .5),
               rep = 1,
               popsize = 20,
               seed = 4)

# Draw 100 bootstrap replicates from
# each of the 4 Q matrices
bs <- Q_bootstrap(matrices = list(A=A,
                                  B=B,
                                  C=C,
                                  D=D),
                  n_replicates = 100,
                  K = 2)

# Access the elements of this list using $.
# For example....

# To look at all 400 bootstrap Q matrix
# replicates:
bs$bootstrap_replicates

# To look at Fst, FstMax, and
# the ratio (Fst/FstMax) for each replicate
bs$statistics

# To look at a plot of the distribution of
# Fst/FstMax for each Q matrix:
bs$plot_violin

# To determine if each of the 4 distibutions of
# Fst/FstMax is significantly different from
# each of the other distributions:
bs$test_pairwise_wilcox


}
